moving event averaging functions to ibl-avg-trial-viz, adding resampling utils to brainbox.task
